Move decoder methods to `PlanetoidDBForm.Decoders.cs`
